--- src/myInfo/MyInfo.css
@@ -1,12 +1,0 @@
-.myInfo {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.infoBtn {
-  width: 350px;
-  margin: 20px 0;
-}
